chore: Update CSS file to include new utility class and highlight text
The CSS file was updated to include a new utility class called "text-balance" which applies a text-wrap property of "balance". Additionally, a new class called "highlightedText" was added to apply the "text-blue-500" utility class. This will be used to highlight specific text on the website.

The commit also includes an update to the npm dependencies to their latest versions.