feat: Add copy functionality to GenerateTemplate component
This commit adds the functionality to copy the highlighted text in the GenerateTemplate component. When the "Copy" button is clicked, the text inside the highlightedTextRef element is copied to the clipboard using the `navigator.clipboard.writeText()` method. If the copy is successful, an alert is shown to indicate that the text has been copied. If there is an error during the copy process, the error is logged to the console.

The code changes include adding an `onClick` event handler to the button in the GenerateTemplate component and using the `highlightedTextRef` to get the text to be copied.